requirements.txt
Django==4.2.5
openai==1.10.0
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.1
